% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodness_of_fit_loglin.R
\name{goodness_of_fit_loglin}
\alias{goodness_of_fit_loglin}
\title{Calculate the Chi^2 and G^2 Statistics}
\usage{
goodness_of_fit_loglin(model, type = "Chisq", ...)
}
\arguments{
\item{model}{a GLM regression model.}

\item{type}{either "Chisq" or "Gsq", which determines the type of goodness of fit test that is ran. Defaults to "Chisq".}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A list with class "\code{ct_goodness_of_fit}" containing the following components:

\code{test} the type of test used.

\code{model} the name of the inputted model.

\code{statistic} The value of the test statistic as determined by the type parameter

\code{df} The number of degrees of freedom.
  This equals the number of combinations for explanatory variables less the number of parameters in the model

\code{p.value} The p-value calculated under a Chi-Squared distribution.
}
\description{
Calculates the goodness of fit test statistics for log linear models
}
\examples{
lung_loglin <-
  lung_cancer \%>\%
  flatten_ct() \%>\%
  group_by_all() \%>\%
  count() \%>\%
  ungroup() \%>\%
  glm(
    n ~ Lung + Smoking + City
    ,family = poisson
    ,data = .
  )

goodness_of_fit_loglin(model = lung_loglin, type = "Chisq")

lung_loglin \%>\%
  goodness_of_fit_loglin(type = "Gsq")

}
