% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_ct_goodness_of_fit.R
\name{print.ct_goodness_of_fit}
\alias{print.ct_goodness_of_fit}
\title{Print method for goodness_of_fit()}
\usage{
\method{print}{ct_goodness_of_fit}(x, ...)
}
\arguments{
\item{x}{A list}
}
\description{
Creates a nice looking output for the goodness_of_fit() function
}
\author{
Scott D Graham, \email{scott.grah95@gmail.com}
}
