% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodness_of_fit.R
\name{goodness_of_fit}
\alias{goodness_of_fit}
\title{Calculate the Chi^2 and G^2 Statistics}
\usage{
goodness_of_fit(model, ..., response, type = "Chisq")
}
\arguments{
\item{model}{a GLM regression model.}

\item{response}{a string with the same name as the response column in the data}

\item{type}{either "Chisq" or "Gsq", which determines the type of goodness of fit test that is ran. Defaults to "Chisq".}
}
\value{
A list with class "\code{ct_goodness_of_fit}" containing the following components:

\code{test} the type of test used.

\code{model} the name of the inputted model.

\code{statistic} The value of the test statistic as determined by the type parameter

\code{df} The number of degrees of freedom.
  This equals the number of combinations for explanatory variables less the number of parameters in the model

\code{p.value} The p-value calculated under a Chi-Squared distribution.
}
\description{
Calculates the goodness of fit test statistics for contingency tables
}
\examples{
lung_cancer_ct <-
  array(
    data =
      c(
        126, 35, 100, 61
        ,908, 497, 688, 807
        ,913, 336, 747, 598
        ,235, 58, 172, 121
        ,402, 121, 308, 215
        ,182, 72, 156, 98
        ,60, 11, 99, 43
       ,104, 21, 89, 36
      )
    ,dim = c(2, 2, 8)
    ,dimnames = list(
      Smoking = c("Y", "N")
      ,Lung = c("Y", "N")
      ,City =
        c(
          "Beij"
          ,"Shan"
          ,"Shen"
          ,"Nanj"
          ,"Harb"
          ,"Zhen"
          ,"Taiy"
          ,"Nanc"
        )
    )
  )

lung_logit <-
  lung_cancer_ct \%>\%
  flatten_ct() \%>\%
  glm(
    Lung ~ City + Smoking
    ,family = binomial
    ,data = .
  )

goodness_of_fit(model = lung_logit, response = "Lung", type = "Gsq")
lung_logit \%>\%
  goodness_of_fit(response = "Lung", type = "Gsq")
lung_cancer_ct \%>\%
  flatten_ct() \%>\%
  glm(
    Lung ~ City + Smoking
    ,family = binomial
    ,data = .
  ) \%>\%
  goodness_of_fit(response = "Lung", type = "Gsq")

}
\author{
Scott D Graham, \email{scott.grah95@gmail.com}
}
