% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodness_of_fit.R
\name{goodness_of_fit}
\alias{goodness_of_fit}
\title{Calculate the Chi^2 and G^2 Statistics}
\usage{
goodness_of_fit(model, ..., response, type = "Chisq")
}
\arguments{
\item{model}{a GLM regression model.}
}
\value{
A numeric value
}
\description{
goodness_of_fit() calculates the goodness of fit test statistics for contingency tables
}
\examples{
lung_cancer_ct <-
  array(
    data =
      c(
        126, 35, 100, 61
        ,908, 497, 688, 807
        ,913, 336, 747, 598
        ,235, 58, 172, 121
        ,402, 121, 308, 215
        ,182, 72, 156, 98
        ,60, 11, 99, 43
       ,104, 21, 89, 36
      )
    ,dim = c(2, 2, 8)
    ,dimnames = list(
      Smoking = c("Y", "N")
      ,Lung = c("Y", "N")
      ,City =
        c(
          "Beij"
          ,"Shan"
          ,"Shen"
          ,"Nanj"
          ,"Harb"
          ,"Zhen"
          ,"Taiy"
          ,"Nanc"
        )
    )
  )

lung_logit <-
  lung_cancer_ct \%>\%
  flatten_ct() \%>\%
  glm(
    Lung ~ City + Smoking
    ,family = binomial
    ,data = .
  )

goodness_of_fit(model = lung_logit, response = "Lung", type = "Gsq")
lung_logit \%>\%
  goodness_of_fit(response = "Lung", type = "Gsq")
lung_cancer_ct \%>\%
  flatten_ct() \%>\%
  glm(
    Lung ~ City + Smoking
    ,family = binomial
    ,data = .
  ) \%>\%
  goodness_of_fit(response = "Lung", type = "Gsq")

}
\author{
Scott D Graham, \email{scott.grah95@gmail.com}
}
